moved test from app to spring module
