+++ pf4j-spring/src/test/java/org/pf4j/spring/SpringPluginTest.java
@@ -1,0 +1,43 @@
+/*
+ * Copyright (C) 2012-present the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.pf4j.spring;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.test.context.junit.jupiter.SpringJUnitConfig;
+
+@SpringJUnitConfig(TestConfiguration.class)
+public class SpringPluginTest {
+    private static final String PLUGIN_ID = "hello-plugin";
+
+    private final SpringPluginManager pluginManager;
+
+    @Autowired
+    public SpringPluginTest(SpringPluginManager pluginManager) {
+        this.pluginManager = pluginManager;
+    }
+
+    @Test
+    public void testStartStopStartRuns() {
+        pluginManager.startPlugin(PLUGIN_ID);
+        pluginManager.getExtensions(PLUGIN_ID);
+        pluginManager.stopPlugin(PLUGIN_ID);
+
+        pluginManager.startPlugin(PLUGIN_ID);
+        pluginManager.getExtensions(PLUGIN_ID);
+        pluginManager.stopPlugin(PLUGIN_ID);
+    }
+}
