+++ pf4j-spring/src/test/resources/log4j.properties
@@ -1,0 +1,5 @@
+log4j.rootLogger=info,stdout
+
+log4j.appender.stdout=org.apache.log4j.ConsoleAppender
+log4j.appender.stdout.layout=org.apache.log4j.PatternLayout
+log4j.appender.stdout.layout.ConversionPattern=%p\t%d{ISO8601}\t%r\t%c\t[%t]\t%m%n
