+++ pom.xml
@@ -49,8 +49,7 @@
         <spring.version>5.2.9.RELEASE</spring.version>
         <slf4j.version>1.7.25</slf4j.version>
 
-        <junit.version>4.12</junit.version>
-        <mockito.version>2.0.28-beta</mockito.version>
+        <junit.version>5.7.0</junit.version>
         <cobertura.version>2.7</cobertura.version>
         <coveralls.version>3.1.0</coveralls.version>
